refactor(routes): inline SetManagement, use suspense query for duplicates
Addresses review feedback: fully inline the sets admin route's page
component instead of a thin wrapper, and switch the duplicate-artists
route to useSuspenseQuery backed by a loader ensureQueryData call.
